% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{package_offline_bundle}
\alias{package_offline_bundle}
\title{download r pkgs}
\usage{
package_offline_bundle(
  cran_pkgs = NULL,
  github_repos = NULL,
  bioc_pkgs = NULL,
  dest_dir,
  bundle_name
)
}
\arguments{
\item{cran_pkgs}{vectors of pkgs installed from R-Cran}

\item{github_repos}{vectors of github_repos from GitHub}

\item{bioc_pkgs}{vectors of pkgs installed from BioConductor}

\item{dest_dir}{fold for saving all pkgs}

\item{bundle_name}{name of the pkgs}
}
\description{
download r pkgs
}
